fix(plugin-npm-cli): fix login with Verdaccio
This commit fixes `yarn npm login` when the remote registry is Verdaccio.

When a user already exists, the registry replies with `409 Conflict`. The official npm client then retrieves the latest user state and inserts a revision, using HTTP basic authentication. This step was missing, and this commits adds it.

The change was tested to work with a private Verdaccio registry. It should now be as reliable as the official npm client.

- Closes yarnpkg/berry#1044
- Closes yarnpkg/berry#1848
- Closes verdaccio/verdaccio#1737