Add second level of --verbose to `workspaces foreach` to permit specifying less logging. (#6034)
**What's the problem this PR addresses?**

Fixes #6033.

**How did you fix it?**

Changes the `-v,--verbose` flag in `workspaces foreach` to be a counter
rather than a boolean. With one level it prints prefixes, with two
levels it additionally prints start/finish/timing information (as was
the behavior before).

The default TTY behavior has not changed (maximum verbosity). The
behavior with a single `--verbose` _has_ changed, to be less verbose.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.