docs: fix links in zero-installs explanation (#6075)
**What's the problem this PR addresses?**

I spent a long time debugging a failing `yarn install --immutable
--immutable-cache --check-cache` in my CI environment.

I was seeing the error `YN0056: PACKAGE@npm:x.x.x: Cache entry required
but missing for PACKAGE@npm:x.x.x` for every package in the cache. (The
issue was clearly not just for packages with native dependencies like in
https://github.com/yarnpkg/berry/issues/4514.)

After trying in multiple environments, I finally realized `.pnp.cjs` was
changing based on the nesting depth of the repo, as `enableGlobalCache`
was not set to `false` in `.yarnrc.yml`. (This occurred after an upgrade
from Yarn v3 to Yarn v4.)

(Including the above to help anyone searching for this issue on GitHub.)

**How did you fix it?**

This PR fixes the broken PnP link and makes the `offline mirror` text a
link as well. Hopefully that will help others to realize that to get
Zero Installs working they'll need to read that section too.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
- [ ] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [ ] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [ ] I will check that all automated PR checks pass before the PR gets
reviewed.