Adds ts-nocheck to the PnP file (#6183)
**What's the problem this PR addresses?**

The PnP file is fairly large; it seems that parsing it is difficult for
TypeScript on projects configured with `checkJs`, even if it doesn't
contain `// @ts-check`. Adding `// @ts-nocheck` to the file seemed to
improve the situation.

**How did you fix it?**

Adds `// @ts-nocheck` to the generated `.pnp.cjs` file.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.