Improve libc presence detection (#6170)
**What's the problem this PR addresses?**

On certain Linux distributions, `/usr/bin/ldd` does not contain the
string "GLIBC", but "libc". This is the case in popular distros such as
openSUSE, Amazon Linux, Fedora, RHEL, Arch, and more. On these
distributions, the current version of yarn falls back to
`process.report.getReport()`, which is known to have performance issues
in some instances.

**How did you fix it?**

Add a check for the presence of the value `libc` in addition to the
existing check for `GLIBC`. This allows more Linux users to benefit from
the fast-path that does not rely on `process.report.getReport()`.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.