Fixes `yarn why -R` (#6251)
## What's the problem this PR addresses?

When using the `-R` flag on a flag with a large amount of workspaces,
`yarn why` may fail to return enough information to answer the question
"why is this package here". This is because of a bug in the
implementation where we decline to print a package in the tree view if:

- We saw it before, or
- It's a workspace (because we'll print it later anyway)

However we check the first condition first and we put the package in the
"seen" list before we check whether it's a workspace. As a result,
workspaces that are dependencies of other workspaces won't have their
dependencies shown.

## How did you fix it?

Flipped the two conditions.

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.