Remove "sherlock" from bug report template
The linked documentation doesn't exist anymore.
The associated website for the tool [1] shows some useful information,
but (IMHO) it's unclear to see how to apply it in a fresh repo.

Additionally, `arcanis` mentioned in a comment here [2] that perhaps it
should be removed since it's not often used and the playground is
broken.

Fixes #5837
Fixes #5950

[1] https://github.com/arcanis/sherlock
[2] https://github.com/yarnpkg/berry/pull/6139#issuecomment-1977286294