Adds support for ESM loaders to the SDKs (#6278)
## What's the problem this PR addresses?

ESM loaders aren't supported by the sdks because we only
`require().setup()` but we don't `register()` the ESM loader hook.

Fixes #6219

## How did you fix it?

- Add support for ESM loaders to the SDK.

- Check whether the `findPnpApi` function could be found in the `module`
builtin. If not, tries to setup the environment accordingly. This is
because in the case of the SDK, the subprocesses have been created
without the `--require .pnp.loader.mjs` flag, so the PnP runtime won't
be setup.

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.