Adds support for cwd in the git protocol (#6362)
## What's the problem this PR addresses?

The `git:` protocol supports cloning workspaces from a given repository,
but only if the project to install is at the root of the repository. It
may not be the case when the project contains libraries for multiple
languages.

## How did you fix it?

A new `cwd` parameter lets users pick a subpath to use when packing the
project. It can be used on its own, or in tandem with the `workspace`
specifier.

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.