test: fix custom certification tests on Windows
The integration tests using custom certificates have been failing for
quite a while on Windows with the error `RequestError: unsuitable certificate purpose`.
This error was caused by the test CA certificate used; it was not
configured to be used as a CA when it was generated. The test script
that generates the CA certificate has been updated to generate it
correctly.

The problem does not seem to be specific to Windows; I was able to
reproduce it locally on macOS as well. I don't yet know why it's
passing on CI for Ubuntu and macOS, but not for Windows. Perhaps some
difference in the images, like the OpenSSL version (they should all be
using OpenSSL v1.1.1, but I've seen reports of Windows having v3
installed as well).