fix(fslib): handle symlinks to zip files (#6603)
This is the same as https://github.com/yarnpkg/berry/pull/5474 but
allows edits by maintainer

**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
Yarn shouldn't care if zips in the cache are symlinks or not. The only
reason it doesn't work is that finding the mount point skips zips if
they're not regular files. This used to be supported until it was
removed for "performance" reasons in
https://github.com/yarnpkg/berry/pull/1474. The bulk of the logic in the
linked PR involves resolving the real path of the symlink but I don't
think that's needed for this?

<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->
Resolves https://github.com/yarnpkg/berry/issues/3514
Closes https://github.com/yarnpkg/berry/pull/5474 (supersedes it)

**How did you fix it?**
I switched the check to use `stat` from `lstat`.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [X] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [X] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: merceyz <merceyz@users.noreply.github.com>