Add tests for --no-default and --no-default --json
Squashed from several commits while figuring out what was going on:

Remove unnecessary replacements

Back in 7835a7d39598e400c374d0cbf815405022bf223f there was no `FILTER`
at the top but now that there is, these settings are not in the output.

Remove confusing cleanupJsonOutput method

This does not support ndjson so in all tests (prior to this PR)
it was failing to parse the JSON and not actually executing
most of the method body.

Now that one of the new tests returns only a single line of JSON
this 'resurrected' this dead broken method causing tests to fail.

Instead use the same cleanup method for all cases, which is what
has been happening anyway (due to the fallback in the `catch`)

Add the remaining to snapshots now the tests are fixed