add a config option to disallow the cache clean command (#6610)
## What's the problem this PR addresses?

At [Datadog](https://github.com/datadog) we are using `yarn` inside a
huge monorepository using the PnP linker and Zero-Installs with all the
dependencies versioned through git.

The issue we have is that people frequently have the tendency to think
that this kind of environment works like others, and that cleaning the
cache is the way to go whenever they encounter an install issue (like
deleting `node_modules`).

This is obviously not the case and we have to tell them to restore their
packages from the git remote when that happens, which is a bit
cumbersome.

_see also: https://github.com/yarnpkg/berry/discussions/6518_

## How did you fix it?

This PR adds a non-breaking option inside the `.yarnrc` file to disallow
running `yarn clean cache`.

cc @arcanis 

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.