Fix error handling when looking for configuration files
When looking for configuration files if reading a candidate throws an exception the exception is not caught and the user gets a cryptic and non-informative message.
e.g. if you happen to have a directory named `.yarnrc.yml` in the root of your project every yarn command will result with this message:
```
Internal Error: EISDIR: illegal operation on a directory, read
Error: EISDIR: illegal operation on a directory, read
```
which doesn't include the file.

This change moves the reading operation into the try clause and handles read errors in the same way parsing errors are handled.